requirements.txt
asgiref==3.5.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==4.0.6
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-environ==0.9.0
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2==2.8.5
psycopg2-binary==2.9.3
pycodestyle==2.9.0
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
whitenoise==5.1.0
